requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
pyOpenSSL==20.0.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
